#214404 Add multiple giftcerts or coupons in micro-cart
